requirements.txt
absl-py==1.1.0
aiohttp==3.8.1
aiosignal==1.2.0
appnope==0.1.3
asttokens==2.0.5
astunparse==1.6.3
async-timeout==4.0.2
attrs==21.4.0
backcall==0.2.0
benepar==0.2.0
blis==0.7.8
cachetools==5.2.0
catalogue==2.0.7
certifi==2022.6.15
charset-normalizer==2.1.0
click==7.1.2
cycler==0.11.0
cymem==2.0.6
datasets==2.2.2
debugpy==1.6.0
decorator==5.1.1
dill==0.3.4
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.3.0/en_core_web_lg-3.3.0-py3-none-any.whl
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.0.0/en_core_web_sm-3.0.0.tar.gz
entrypoints==0.4
executing==0.8.3
filelock==3.7.1
flatbuffers==2.0
fonttools==4.33.3
frozenlist==1.3.0
fsspec==2022.5.0
gast==0.5.3
google-auth==2.9.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.47.0
h5py==3.7.0
huggingface-hub==0.8.1
idna==3.3
imageio==2.19.3
importlib-metadata==4.12.0
ipykernel==6.15.0
ipython==8.4.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jupyter-client==7.3.4
jupyter-core==4.10.0
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.3
langcodes==3.3.0
libclang==14.0.1
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
multidict==6.0.2
multiprocess==0.70.12.2
murmurhash==1.0.7
nest-asyncio==1.5.5
networkx==2.8.4
nltk==3.7
numpy==1.22.4
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.3
parso==0.8.3
pathy==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
preshed==3.0.6
prompt-toolkit==3.0.30
protobuf==3.19.4
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.8.2
pyfakefs==4.5.6
Pygments==2.12.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==23.2.0
regex==2022.6.2
requests==2.28.1
requests-oauthlib==1.3.1
responses==0.18.0
rsa==4.8
sacremoses==0.0.53
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
sentence-transformers==2.2.2
sentencepiece==0.1.96
six==1.16.0
sklearn==0.0
smart-open==5.2.1
spacy==3.0.8
spacy-legacy==3.0.9
spacy-loggers==1.0.2
srsly==2.4.3
stack-data==0.3.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.2
tensorflow-estimator==2.8.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
thinc==8.0.17
threadpoolctl==3.1.0
tifffile==2022.5.4
tokenizers==0.12.1
torch==1.11.0
torch-struct==0.5
torchvision==0.12.0
tornado==6.2
tqdm==4.64.0
traitlets==5.3.0
transformers==4.17.0
typer==0.3.2
typing_extensions==4.3.0
urllib3==1.26.9
wasabi==0.9.1
wcwidth==0.2.5
Werkzeug==2.1.2
wrapt==1.14.1
xxhash==3.0.0
yarl==1.7.2
zipp==3.8.0
